From be45a36a5d6ae99eb1c2865aa4b4f5fd9b20b1ed Mon Sep 17 00:00:00 2001 [PATCH] x86/msr: Blacklist various MSRs which guests definitely
 shouldn't be using

The main purpose is to blacklist the Intel Resource Director Technology MSRs.
We do not yet virtualise support for guests, but Linux has been observed to
probe for these MSRs without checking CPUID first.

The architecturally inaccessable ranges don't need to fall back into the
legacy ranges, because they are not going to eventually evaluate as
accessible.

The Silicon Debug interface will probably never be virtualised for guests, but
doesn't want to leak through from real hardware.  SGX isn't yet virtualised,
but likely will be in the future.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
[dropped the blacklist bits]
Signed-off-by: Sergey Dyasli <sergey.dyasli@citrix.com>