x86: Prioritize low memory size from Multiboot
Previously, Xen used information from the BDA to detect the amount of
available low memory. This does not work on some scenarios such as
Coreboot, or when booting from Kexec on a UEFI system without CSM.

Prioritize the information supplied by Multiboot instead. If this is not
available, fall back to the old BDA method.