livepatch: Ignore duplicate new symbols
For XenServer, we build cumulative live patches and replace so there is
only ever a single active one. This may cause duplicate new symbol
errors. Because of the lookup order (hypervisor first) and the fact that
live patches will only ever be built against the hypervisor (and not
another live patch), it should be safe to ignore these errors.